Allow to specify when to send raw data in POST/PUT
In e5534bc the post and put methods were modified to only set
the Content-Type when data is a dict. This broke existing users
where data is sent as a json string, because it prevented the
Content-Type header from being added to the request.

Introduce a new `raw` parameter to allow callers to explicitly
decide when a request should be sent with raw data.

Change-Id: I4979c39b1d48a77fc39b6ceece1c84a887f02bdb